backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
httpx
pydantic-settings
python-dotenv
apscheduler
tenacity
pandas
numpy
statsmodels
